feat: Move Notification & Permission & Properties Repositories to new folder structure
