+++ src/script/components/AppContainer/AppContainer.tsx
@@ -27,10 +27,10 @@ import {StyledApp, THEME_ID} from '@wireapp/react-ui-kit';
 import {DetachedCallingCell} from 'Components/calling/DetachedCallingCell';
 import {PrimaryModalComponent} from 'Components/Modals/PrimaryModal/PrimaryModal';
 import {QualityFeedbackModal} from 'Components/Modals/QualityFeedbackModal';
+import {useAppSoftLock} from 'Hooks/useAppSoftLock';
+import {useSingleInstance} from 'Hooks/useSingleInstance';
+import {PROPERTIES_TYPE} from 'Repositories/properties/PropertiesType';
 import {SIGN_OUT_REASON} from 'src/script/auth/SignOutReason';
-import {useAppSoftLock} from 'src/script/hooks/useAppSoftLock';
-import {useSingleInstance} from 'src/script/hooks/useSingleInstance';
-import {PROPERTIES_TYPE} from 'src/script/properties/PropertiesType';
 import {isDetachedCallingFeatureEnabled} from 'Util/isDetachedCallingFeatureEnabled';
 
 import {useAccentColor} from './hooks/useAccentColor';
