+++ src/script/components/Conversation/Conversation.tsx
@@ -36,8 +36,8 @@ import {TitleBar} from 'Components/TitleBar';
 import {CallState} from 'Repositories/calling/CallState';
 import {ConversationState} from 'Repositories/conversation/ConversationState';
 import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
+import {PROPERTIES_TYPE} from 'Repositories/properties/PropertiesType';
 import {Config} from 'src/script/Config';
-import {PROPERTIES_TYPE} from 'src/script/properties/PropertiesType';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {isLastReceivedMessage} from 'Util/conversationMessages';
 import {allowsAllFiles, getFileExtensionOrName, hasAllowedExtension} from 'Util/FileTypeUtil';
