+++ src/script/hooks/useAppSoftLock.ts
@@ -20,9 +20,9 @@
 import {useCallback, useEffect, useState} from 'react';
 
 import {CallingRepository} from 'Repositories/calling/CallingRepository';
+import {NotificationRepository} from 'Repositories/notification/NotificationRepository';
 
 import {E2EIHandler, E2EIDeviceStatus} from '../E2EIdentity';
-import {NotificationRepository} from '../notification/NotificationRepository';
 
 export function useAppSoftLock(callingRepository: CallingRepository, notificationRepository: NotificationRepository) {
   const [softLockEnabled, setSoftLockEnabled] = useState(false);
