+++ src/script/main/app.ts
@@ -62,6 +62,11 @@ import {GiphyService} from 'Repositories/giphy/GiphyService';
 import {IntegrationRepository} from 'Repositories/integration/IntegrationRepository';
 import {IntegrationService} from 'Repositories/integration/IntegrationService';
 import {MediaRepository} from 'Repositories/media/MediaRepository';
+import {NotificationRepository} from 'Repositories/notification/NotificationRepository';
+import {PreferenceNotificationRepository} from 'Repositories/notification/PreferenceNotificationRepository';
+import {PermissionRepository} from 'Repositories/permission/PermissionRepository';
+import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
+import {PropertiesService} from 'Repositories/properties/PropertiesService';
 import {initializeDataDog} from 'Util/DataDog';
 import {DebugUtil} from 'Util/DebugUtil';
 import {Environment} from 'Util/Environment';
@@ -87,13 +92,8 @@ import {externalUrl} from '../externalRoute';
 import {startNewVersionPolling} from '../lifecycle/newVersionHandler';
 import {initMLSGroupConversations, initialiseSelfAndTeamConversations} from '../mls';
 import {joinConversationsAfterMigrationFinalisation} from '../mls/MLSMigration/migrationFinaliser';
-import {NotificationRepository} from '../notification/NotificationRepository';
-import {PreferenceNotificationRepository} from '../notification/PreferenceNotificationRepository';
 import {configureDownloadPath} from '../page/components/FeatureConfigChange/FeatureConfigChangeHandler/Features/downloadPath';
 import {configureE2EI} from '../page/components/FeatureConfigChange/FeatureConfigChangeHandler/Features/E2EIdentity';
-import {PermissionRepository} from '../permission/PermissionRepository';
-import {PropertiesRepository} from '../properties/PropertiesRepository';
-import {PropertiesService} from '../properties/PropertiesService';
 import {SearchRepository} from '../search/SearchRepository';
 import {SelfRepository} from '../self/SelfRepository';
 import {SelfService} from '../self/SelfService';
