+++ src/script/page/LeftSidebar/panels/Conversations/ConversationCallingView/ConversationCallingView.tsx
@@ -19,8 +19,8 @@
 
 import {CallingCell} from 'Components/calling/CallingCell';
 import {Call} from 'Repositories/calling/Call';
-import {PropertiesRepository} from 'src/script/properties/PropertiesRepository';
-import {PROPERTIES_TYPE} from 'src/script/properties/PropertiesType';
+import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
+import {PROPERTIES_TYPE} from 'Repositories/properties/PropertiesType';
 import {ListViewModel} from 'src/script/view_model/ListViewModel';
 
 type ConversationCallingViewProps = {
