+++ src/script/page/MainContent/panels/preferences/AVPreferences.tsx
@@ -21,6 +21,7 @@ import {useInitializeMediaDevices} from 'Hooks/useInitializeMediaDevices';
 import type {CallingRepository} from 'Repositories/calling/CallingRepository';
 import {MediaDeviceType} from 'Repositories/media/MediaDeviceType';
 import type {MediaRepository} from 'Repositories/media/MediaRepository';
+import type {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
 
@@ -30,8 +31,6 @@ import {CameraPreferences} from './avPreferences/CameraPreferences';
 import {MicrophonePreferences} from './avPreferences/MicrophonePreferences';
 import {SaveCallLogs} from './avPreferences/SaveCallLogs';
 import {PreferencesPage} from './components/PreferencesPage';
-
-import type {PropertiesRepository} from '../../../../properties/PropertiesRepository';
 
 interface AVPreferencesProps {
   callingRepository: CallingRepository;
