+++ src/script/page/MainContent/panels/preferences/AccountPreferences.tsx
@@ -29,6 +29,7 @@ import {useEnrichedFields} from 'Components/panel/EnrichedFields';
 import {ClientRepository} from 'Repositories/client';
 import {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
 import {ConversationState} from 'Repositories/conversation/ConversationState';
+import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
 import {ContentState} from 'src/script/page/useAppState';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {t} from 'Util/LocalizerUtil';
@@ -51,7 +52,6 @@ import {PreferencesSection} from './components/PreferencesSection';
 
 import {Config} from '../../../../Config';
 import {User} from '../../../../entity/User';
-import {PropertiesRepository} from '../../../../properties/PropertiesRepository';
 import {TeamState} from '../../../../team/TeamState';
 import {RichProfileRepository} from '../../../../user/RichProfileRepository';
 import type {UserRepository} from '../../../../user/UserRepository';
