+++ src/script/page/MainContent/panels/preferences/avPreferences/CallOptions.tsx
@@ -26,11 +26,11 @@ import {Checkbox, CheckboxLabel} from '@wireapp/react-ui-kit';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import type {MediaConstraintsHandler} from 'Repositories/media/MediaConstraintsHandler';
+import type {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
+import {PROPERTIES_TYPE} from 'Repositories/properties/PropertiesType';
 import {t} from 'Util/LocalizerUtil';
 
 import {Config} from '../../../../../Config';
-import type {PropertiesRepository} from '../../../../../properties/PropertiesRepository';
-import {PROPERTIES_TYPE} from '../../../../../properties/PropertiesType';
 import {PreferencesSection} from '../components/PreferencesSection';
 
 interface CallOptionsProps {
