+++ src/script/repositories/conversation/ConversationLabelRepository.ts
@@ -24,6 +24,7 @@ import ko from 'knockout';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {PrimaryModal} from 'Components/Modals/PrimaryModal';
+import type {PropertiesService} from 'Repositories/properties/PropertiesService';
 import {SidebarTabs, useSidebarStore} from 'src/script/page/LeftSidebar/panels/Conversations/useSidebarStore';
 import {t} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
@@ -31,7 +32,6 @@ import {TypedEventTarget} from 'Util/TypedEventTarget';
 import {createUuid} from 'Util/uuid';
 
 import type {Conversation} from '../../entity/Conversation';
-import type {PropertiesService} from '../../properties/PropertiesService';
 
 export enum LabelType {
   Custom = 0,
