+++ src/script/repositories/conversation/ConversationRepository.ts
@@ -71,6 +71,7 @@ import {EventRepository} from 'Repositories/event/EventRepository';
 import {EventService} from 'Repositories/event/EventService';
 import {EventSource} from 'Repositories/event/EventSource';
 import {NOTIFICATION_HANDLING_STATE} from 'Repositories/event/NotificationHandlingState';
+import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
 import {getNextItem} from 'Util/ArrayUtil';
 import {allowsAllFiles, getFileExtensionOrName, isAllowedFile} from 'Util/FileTypeUtil';
 import {replaceLink, t} from 'Util/LocalizerUtil';
@@ -152,7 +153,6 @@ import * as LegalHoldEvaluator from '../../legal-hold/LegalHoldEvaluator';
 import {MessageCategory} from '../../message/MessageCategory';
 import {SystemMessageType} from '../../message/SystemMessageType';
 import {addOtherSelfClientsToMLSConversation} from '../../mls';
-import {PropertiesRepository} from '../../properties/PropertiesRepository';
 import {SelfRepository} from '../../self/SelfRepository';
 import {Core} from '../../service/CoreSingleton';
 import type {EventRecord} from '../../storage';
