+++ src/script/repositories/media/MediaRepository.ts
@@ -17,11 +17,11 @@
  *
  */
 
+import type {PermissionRepository} from 'Repositories/permission/PermissionRepository';
+
 import {MediaConstraintsHandler} from './MediaConstraintsHandler';
 import {MediaDevicesHandler} from './MediaDevicesHandler';
 import {MediaStreamHandler} from './MediaStreamHandler';
-
-import type {PermissionRepository} from '../../permission/PermissionRepository';
 
 export class MediaRepository {
   public constraintsHandler: MediaConstraintsHandler;
