+++ src/script/repositories/notification/NotificationRepository.ts
@@ -37,6 +37,9 @@ import type {ConnectionEntity} from 'Repositories/connection/ConnectionEntity';
 import {ConversationEphemeralHandler} from 'Repositories/conversation/ConversationEphemeralHandler';
 import type {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
 import {ConversationState} from 'Repositories/conversation/ConversationState';
+import type {PermissionRepository} from 'Repositories/permission/PermissionRepository';
+import {PermissionStatusState} from 'Repositories/permission/PermissionStatusState';
+import {PermissionType} from 'Repositories/permission/PermissionType';
 import {Declension, t, getUserName} from 'Util/LocalizerUtil';
 import {getLogger, Logger} from 'Util/Logger';
 import {getRenderedTextContent} from 'Util/messageRenderer';
@@ -46,24 +49,21 @@ import {ValidationUtilError} from 'Util/ValidationUtil';
 
 import {PermissionState} from './PermissionState';
 
-import type {Conversation} from '../entity/Conversation';
-import type {CallMessage} from '../entity/message/CallMessage';
-import type {ContentMessage} from '../entity/message/ContentMessage';
-import type {DeleteConversationMessage} from '../entity/message/DeleteConversationMessage';
-import type {MemberMessage} from '../entity/message/MemberMessage';
-import type {Message} from '../entity/message/Message';
-import type {MessageTimerUpdateMessage} from '../entity/message/MessageTimerUpdateMessage';
-import type {RenameMessage} from '../entity/message/RenameMessage';
-import type {SystemMessage} from '../entity/message/SystemMessage';
-import type {User} from '../entity/User';
-import {SuperType} from '../message/SuperType';
-import {SystemMessageType} from '../message/SystemMessageType';
-import {ContentState, useAppState} from '../page/useAppState';
-import type {PermissionRepository} from '../permission/PermissionRepository';
-import {PermissionStatusState} from '../permission/PermissionStatusState';
-import {PermissionType} from '../permission/PermissionType';
-import {UserState} from '../user/UserState';
-import {Warnings} from '../view_model/WarningsContainer';
+import type {Conversation} from '../../entity/Conversation';
+import type {CallMessage} from '../../entity/message/CallMessage';
+import type {ContentMessage} from '../../entity/message/ContentMessage';
+import type {DeleteConversationMessage} from '../../entity/message/DeleteConversationMessage';
+import type {MemberMessage} from '../../entity/message/MemberMessage';
+import type {Message} from '../../entity/message/Message';
+import type {MessageTimerUpdateMessage} from '../../entity/message/MessageTimerUpdateMessage';
+import type {RenameMessage} from '../../entity/message/RenameMessage';
+import type {SystemMessage} from '../../entity/message/SystemMessage';
+import type {User} from '../../entity/User';
+import {SuperType} from '../../message/SuperType';
+import {SystemMessageType} from '../../message/SystemMessageType';
+import {ContentState, useAppState} from '../../page/useAppState';
+import {UserState} from '../../user/UserState';
+import {Warnings} from '../../view_model/WarningsContainer';
 
 type NotificationData = {conversationId?: QualifiedId; messageId?: string; messageType: string};
 interface NotificationContent {
