+++ src/script/repositories/notification/PermissionState.ts
