+++ src/script/repositories/permission/PermissionStatusState.ts
