+++ src/script/repositories/permission/PermissionType.ts
