+++ src/script/repositories/properties/PropertiesType.ts
