+++ src/script/view_model/ListViewModel.ts
@@ -28,6 +28,7 @@ import {PrimaryModal, usePrimaryModalState} from 'Components/Modals/PrimaryModal
 import type {CallingRepository} from 'Repositories/calling/CallingRepository';
 import type {ConversationRepository} from 'Repositories/conversation/ConversationRepository';
 import {ConversationState} from 'Repositories/conversation/ConversationState';
+import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
 import {iterateItem} from 'Util/ArrayUtil';
 import {isEscapeKey} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
@@ -43,7 +44,6 @@ import {SidebarTabs, useSidebarStore} from '../page/LeftSidebar/panels/Conversat
 import {PanelState} from '../page/RightSidebar';
 import {useAppMainState} from '../page/state';
 import {ContentState, ListState, useAppState} from '../page/useAppState';
-import {PropertiesRepository} from '../properties/PropertiesRepository';
 import {SearchRepository} from '../search/SearchRepository';
 import type {TeamRepository} from '../team/TeamRepository';
 import {TeamState} from '../team/TeamState';
