+++ test/helper/TestFactory.js
@@ -46,10 +46,10 @@ import {EventRepository} from 'Repositories/event/EventRepository';
 import {EventService} from 'Repositories/event/EventService';
 import {NotificationService} from 'Repositories/event/NotificationService';
 import {MediaRepository} from 'Repositories/media/MediaRepository';
+import {PermissionRepository} from 'Repositories/permission/PermissionRepository';
+import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
+import {PropertiesService} from 'Repositories/properties/PropertiesService';
 import {User} from 'src/script/entity/User';
-import {PermissionRepository} from 'src/script/permission/PermissionRepository';
-import {PropertiesRepository} from 'src/script/properties/PropertiesRepository';
-import {PropertiesService} from 'src/script/properties/PropertiesService';
 import {SearchRepository} from 'src/script/search/SearchRepository';
 import {SelfService} from 'src/script/self/SelfService';
 import {Core} from 'src/script/service/CoreSingleton';
