+++ test/unit_tests/properties/PropertiesRepositorySpec.js
@@ -17,8 +17,8 @@
  *
  */
 
-import {PropertiesRepository} from 'src/script/properties/PropertiesRepository';
-import {PropertiesService} from 'src/script/properties/PropertiesService';
+import {PropertiesRepository} from 'Repositories/properties/PropertiesRepository';
+import {PropertiesService} from 'Repositories/properties/PropertiesService';
 import {SelfService} from 'src/script/self/SelfService';
 
 describe('PropertiesRepository', () => {
